refactor(behat): remove FoundryCallFilter and FoundryTableNode
The table normalization (object references, dates, enums) moves from a Behat
call filter to a TableParametersNormalizer service invoked by the built-in
steps themselves, which already receive the factory short name. This drops the
fragile requirement that overridden steps name their argument "$factoryShortName",
and the reflection hack FoundryTableNode needed to carry objects in table cells.

Console output is unchanged: Behat printers always render the source Gherkin
table, never the transformed argument.
